requirements.txt
appdirs==1.4.3
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bcrypt==3.1.7
blinker==1.8.2
Brotli==1.1.0
cachetools==4.0.0
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.0
distro==1.4.0
Django==4.2.23
django-admin-adminlte==1.0.5
django-import-export==4.1.1
django-tenants==3.7.0
entrypoints==0.3
et_xmlfile==2.0.0
fasteners==0.14.1
filelock==3.0.12
Flask==3.0.0
fonttools==4.58.4
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
monotonic==1.6
mysqlclient==2.2.5
openpyxl==3.1.5
oscrypto==1.3.0
pdf2image==1.17.0
pillow==10.4.0
poppler-utils==0.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.0
reportlab==3.6.13
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tablib==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.11.0
xhtml2pdf==0.2.11
zopfli==0.2.3.post1

# Nuevas dependencias para el dashboard
twilio==9.0.0
django-filter==24.1
pandas==2.2.0